+++ CHANGELOG.md
@@ -14,11 +14,14 @@ All notable changes to this project will be documented in this file.
 - Add ping endpoints for "log" events
 - Add the "Badges" page in docs
 - Add support for multiple recipients in incoming email (#669)
+- Upgrade to fido2 1.0.0, requests 2.28.1, segno 1.5.2
 
 ### Bug Fixes
 - Fix the display of ignored pings with non-zero exit status
 - Fix a race condition in the "Change Email" flow
 - Fix grouping and sorting in the text version of the report/nag emails (#679)
+- Fix the update_timeout view to record more information for downtime bookkeeping
+- Fix the checks list to preserve selected filters when adding/updating checks (#684)
 
 ## v2.2.1 - 2022-06-13